orbit_shared

Module addressable

Structs

AddressableInvocation

An invocation of a method on an addressable

AddressableInvocationArgument

AddressableLease

A lease for an addressable indicating its current location and lifetime

AddressableReference

Reference to an addressable (type + key)

NamespacedAddressableReference

Namespaced reference to an addressable

Enums

Key

A key that uniquely identifies an addressable within its type

Traits

ActorWithInt32Key

Trait for addressables with i32 keys

ActorWithInt64Key

Trait for addressables with i64 keys

ActorWithNoKey

Trait for addressables with no key

ActorWithStringKey

Trait for addressables with string keys

Addressable

Trait marker for addressable actors

Type Aliases

AddressableInvocationArguments

Arguments for an addressable invocation Each argument is a tuple of (value, type_name) for serialization purposes

AddressableType