orbit_shared::saga

Trait SagaStep

pub trait SagaStep: Send + Sync {
    // Required methods
    fn execute<'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: &'life1 SagaContext,
    ) -> Pin<Box<dyn Future<Output = OrbitResult<StepResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn compensate<'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: &'life1 SagaContext,
    ) -> Pin<Box<dyn Future<Output = OrbitResult<StepResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn metadata(&self) -> &SagaStepMetadata;

    // Provided methods
    fn should_execute<'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: &'life1 SagaContext,
    ) -> Pin<Box<dyn Future<Output = OrbitResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn prepare<'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: &'life1 SagaContext,
    ) -> Pin<Box<dyn Future<Output = OrbitResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn cleanup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: &'life1 SagaContext,
    ) -> Pin<Box<dyn Future<Output = OrbitResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

A saga step definition with forward and compensation actions

Required Methods

fn execute<'life0, 'life1, 'async_trait>( &'life0 self, context: &'life1 SagaContext, ) -> Pin<Box<dyn Future<Output = OrbitResult<StepResult>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Execute the forward action of this step

fn compensate<'life0, 'life1, 'async_trait>( &'life0 self, context: &'life1 SagaContext, ) -> Pin<Box<dyn Future<Output = OrbitResult<StepResult>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Execute the compensation action for this step

fn metadata(&self) -> &SagaStepMetadata

Get metadata for this step

Provided Methods

fn should_execute<'life0, 'life1, 'async_trait>( &'life0 self, context: &'life1 SagaContext, ) -> Pin<Box<dyn Future<Output = OrbitResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Check if this step should be executed based on context

fn prepare<'life0, 'life1, 'async_trait>( &'life0 self, context: &'life1 SagaContext, ) -> Pin<Box<dyn Future<Output = OrbitResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Prepare the step for execution (resource allocation, validation, etc.)

fn cleanup<'life0, 'life1, 'async_trait>( &'life0 self, context: &'life1 SagaContext, ) -> Pin<Box<dyn Future<Output = OrbitResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Clean up after step execution (resource deallocation, etc.)